reduce kv usage
