+++ src/tarkov-data-manager/jobs/index.js
@@ -27,17 +27,14 @@ const defaultJobs = {
     'update-traders': '*/5 * * * *',
     'update-barters': '*/5 * * * *',
     'update-crafts': '1-59/5 * * * *',
-    'update-hideout-legacy': '40 * * * *',
-    'update-quests': '42 * * * *',
     'update-existing-bases': '4-59/5 * * * *',
     'game-data': '*/10 * * * *',
     'update-historical-prices': '5-59/15 * * * *',
     'update-trader-prices': '25 9,21 * * *',
     'clear-checkouts': '5,35 * * * *',
     'verify-wiki': '5 9 * * *',
-    'update-ammo': '*/10 * * * *',
     'update-hideout': '2-59/10 * * * *',
-    'update-quests-new': '6-59/10 * * * *',
+    'update-quests': '6-59/10 * * * *',
     'update-presets': '*/10 * * * *',
     'update-maps': '*/20 * * * *',
     // Too much memory :'(
