+++ src/tarkov-data-manager/jobs/update-crafts.js
@@ -130,7 +130,7 @@ module.exports = async function() {
         }
         logger.log(`Processed ${Object.keys(json).length} crafts`);
 
-        const response = await cloudflare('craft_data', 'PUT', JSON.stringify(crafts)).catch(error => {
+        const response = await cloudflare.put('craft_data', JSON.stringify(crafts)).catch(error => {
             logger.error(error);
             return {success: false, errors: [], messages: []};
         });
